move bone categories to json schema
