+++ Ktisis/Data/Schema/BoneCategories.json
@@ -1,0 +1,423 @@
+[
+	{
+		"Name": "Body",
+		"Bones": [
+			"n_hara",
+			"j_kosi",
+			"j_sebo_a",
+			"j_sebo_b",
+			"j_sebo_c",
+			"j_kubi"
+		],
+		"SubCategories": [
+			{
+				"Name": "Head",
+				"Bones": [
+					"j_kao",
+					"j_ago"
+				],
+				"SubCategories": [
+					{
+						"Name": "Hair",
+						"Bones": [
+							"j_kami_a",
+							"j_kami_b",
+							"j_kami_f_l",
+							"j_kami_f_r"
+						]
+					},
+					{
+						"Name": "Ears",
+						"Bones": [
+							"j_mimi_l",
+							"j_mimi_r",
+							"j_zera_a_l",
+							"j_zera_a_r",
+							"j_zera_b_l",
+							"j_zera_b_r",
+							"j_zerb_a_l",
+							"j_zerb_a_r",
+							"j_zerb_b_l",
+							"j_zerb_b_r",
+							"j_zerc_a_l",
+							"j_zerc_a_r",
+							"j_zerc_b_l",
+							"j_zerc_b_r",
+							"j_zerd_a_l",
+							"j_zerd_a_r",
+							"j_zerd_b_l",
+							"j_zerd_b_r"
+						]
+					},
+					{
+						"Name": "Face",
+						"Bones": [
+							"j_f_memoto",
+							"j_f_hana"
+						],
+						"SubCategories": [
+							{
+								"Name": "Brow",
+								"Bones": [
+									"j_f_miken_l",
+									"j_f_miken_r"
+								]
+							},
+							{
+								"Name": "Eyes",
+								"Bones": [
+									"j_f_eye_l",
+									"j_f_eye_r"
+								],
+								"SubCategories": [
+									{
+										"Name": "Eyelids",
+										"Bones": [
+											"j_f_dmab_l",
+											"j_f_dmab_r",
+											"j_f_umab_l",
+											"j_f_umab_r"
+										]
+									},
+									{
+										"Name": "Eyebrows",
+										"Bones": [
+											"j_f_mayu_l",
+											"j_f_mayu_r"
+										]
+									}
+								]
+							},
+							{
+								"Name": "Mouth",
+								"Bones": [
+									"j_f_lip_l",
+									"j_f_lip_r"
+								],
+								"SubCategories": [
+									{
+										"Name": "Upper Lip",
+										"Bones": [
+											"j_f_ulip_a",
+											"j_f_ulip_b"
+										]
+									},
+									{
+										"Name": "Lower Lip",
+										"Bones": [
+											"j_f_dlip_a",
+											"j_f_dlip_b"
+										]
+									}
+								]
+							},
+							{
+								"Name": "Cheeks",
+								"Bones": [
+									"j_f_hoho_l",
+									"j_f_hoho_r"
+								]
+							}
+						]
+					}
+				]
+			},
+			{
+				"Name": "Arms",
+				"SubCategories": [
+					{
+						"Name": "Left Arm",
+						"Bones": [
+							"j_sako_l",
+							"j_ude_a_l",
+							"j_ude_b_l",
+							"n_hkata_l",
+							"n_hhiji_l"
+						],
+						"SubCategories": [
+							{
+								"Name": "Left Hand",
+								"Bones": [
+									"j_hito_a_l",
+									"j_ko_a_l",
+									"j_kusu_a_l",
+									"j_naka_a_l",
+									"j_oya_a_l",
+									"j_hito_b_l",
+									"j_ko_b_l",
+									"j_kusu_b_l",
+									"j_naka_b_l",
+									"j_oya_b_l",
+									"j_te_l",
+									"n_hte_l"
+								],
+								"SubCategories": [
+									{
+										"Name": "IVCS Left Hand",
+										"Bones": [
+											"iv_ko_c_l",
+											"iv_kusu_c_l",
+											"iv_naka_c_l",
+											"iv_hito_c_l"
+										]
+									}
+								]
+							}
+						]
+					},
+					{
+						"Name": "Right Arm",
+						"Bones": [
+							"j_sako_r",
+							"j_ude_a_r",
+							"j_ude_b_r",
+							"n_hkata_r",
+							"n_hhiji_r"
+						],
+						"SubCategories": [
+							{
+								"Name": "Right Hand",
+								"Bones": [
+									"j_hito_a_r",
+									"j_ko_a_r",
+									"j_kusu_a_r",
+									"j_naka_a_r",
+									"j_oya_a_r",
+									"j_hito_b_r",
+									"j_ko_b_r",
+									"j_kusu_b_r",
+									"j_naka_b_r",
+									"j_oya_b_r",
+									"j_te_r",
+									"n_hte_r"
+								],
+								"SubCategories": [
+									{
+										"Name": "IVCS Right Hand",
+										"Bones": [
+											"iv_ko_c_r",
+											"iv_kusu_c_r",
+											"iv_naka_c_r",
+											"iv_hito_c_r"
+										]
+									}
+								]
+							}
+						]
+					}
+				]
+			},
+			{
+				"Name": "Breasts",
+				"Bones": [
+					"j_mune_l",
+					"j_mune_r"
+				],
+				"SubCategories": [
+					{
+						"Name": "IVCS Breasts",
+						"Bones": [
+							"iv_c_mune_l",
+							"iv_c_mune_r"
+						]
+					}
+				]
+			},
+			{
+				"Name": "Tail",
+				"Bones": [
+					"n_sippo_a",
+					"n_sippo_b",
+					"n_sippo_c",
+					"n_sippo_d",
+					"n_sippo_e"
+				]
+			},
+			{
+				"Name": "Legs",
+				"SubCategories": [
+					{
+						"Name": "Left Leg",
+						"Bones": [
+							"j_asi_a_l",
+							"j_asi_b_l",
+							"j_asi_c_l"
+						],
+						"SubCategories": [
+							{
+								"Name": "Left Foot",
+								"Bones": [
+									"j_asi_d_l",
+									"j_asi_e_l"
+								],
+								"SubCategories": [
+									{
+										"Name": "IVCS Left Foot",
+										"Bones": [
+											"iv_asi_oya_a_l",
+											"iv_asi_oya_b_l",
+											"iv_asi_hito_a_l",
+											"iv_asi_hito_b_l",
+											"iv_asi_naka_a_l",
+											"iv_asi_naka_b_l",
+											"iv_asi_kusu_a_l",
+											"iv_asi_kusu_b_l",
+											"iv_asi_ko_a_l",
+											"iv_asi_ko_b_l"
+										]
+									}
+								]
+							}
+						]
+					},
+					{
+						"Name": "Right Leg",
+						"Bones": [
+							"j_asi_a_r",
+							"j_asi_b_r",
+							"j_asi_c_r"
+						],
+						"SubCategories": [
+							{
+								"Name": "Right Foot",
+								"Bones": [
+									"j_asi_d_r",
+									"j_asi_e_r"
+								],
+								"SubCategories": [
+									{
+										"Name": "IVCS Right Foot",
+										"Bones": [
+											"iv_asi_oya_a_r",
+											"iv_asi_oya_b_r",
+											"iv_asi_hito_a_r",
+											"iv_asi_hito_b_r",
+											"iv_asi_naka_a_r",
+											"iv_asi_naka_b_r",
+											"iv_asi_kusu_a_r",
+											"iv_asi_kusu_b_r",
+											"iv_asi_ko_a_r",
+											"iv_asi_ko_b_r"
+										]
+									}
+								]
+							}
+						]
+					}
+				]
+			},
+			{
+				"Name": "IVCS Body",
+				"Bones": [
+					"iv_nitoukin_l",
+					"iv_nitoukin_r"
+				]
+			},
+			{
+				"Name": "IVCS Penis",
+				"Bones": [
+					"iv_kougan_l",
+					"iv_kougan_r",
+					"iv_ochinko_a",
+					"iv_ochinko_b",
+					"iv_ochinko_c",
+					"iv_ochinko_d",
+					"iv_ochinko_e",
+					"iv_ochinko_f"
+				],
+				"IsNsfw": true
+			},
+			{
+				"Name": "IVCS Vagina",
+				"Bones": [
+					"iv_omanko",
+					"iv_kuritto",
+					"iv_inshin_l",
+					"iv_inshin_r"
+				],
+				"IsNsfw": true
+			},
+			{
+				"Name": "IVCS Buttocks",
+				"Bones": [
+					"iv_koumon",
+					"iv_koumon_l",
+					"iv_koumon_r",
+					"iv_shiri_l",
+					"iv_shiri_r"
+				],
+				"IsNsfw": true
+			}
+		]
+	},
+	{
+		"Name": "Clothing",
+		"Bones": [
+			"n_hizasoubi_l",
+			"n_hizasoubi_r",
+			"n_kataarmor_l",
+			"n_kataarmor_r",
+			"n_hijisoubi_l",
+			"n_hijisoubi_r",
+			"j_ex_top_a_r",
+			"j_ex_top_a_l",
+			"j_ex_top_b_r",
+			"j_ex_top_b_l",
+			"j_ex_met_a",
+			"j_ex_met_b",
+			"j_ex_met_c",
+			"j_ex_met_d",
+			"j_ex_met_va",
+			"j_ex_met_vb"
+		],
+		"SubCategories": [
+			{
+				"Name": "Cloth",
+				"Bones": [
+					"j_sk_b_b_l",
+					"j_sk_b_b_r",
+					"j_sk_f_b_l",
+					"j_sk_f_b_r",
+					"j_sk_s_b_l",
+					"j_sk_s_b_r",
+					"j_sk_b_a_l",
+					"j_sk_b_a_r",
+					"j_sk_f_a_l",
+					"j_sk_f_a_r",
+					"j_sk_s_a_l",
+					"j_sk_s_a_r",
+					"j_sk_b_c_l",
+					"j_sk_b_c_r",
+					"j_sk_f_c_l",
+					"j_sk_f_c_r",
+					"j_sk_s_c_l",
+					"j_sk_s_c_r"
+				]
+			},
+			{
+				"Name": "Earring",
+				"Bones": [
+					"n_ear_a_l",
+					"n_ear_a_r",
+					"n_ear_b_l",
+					"n_ear_b_r"
+				]
+			}
+		]
+	},
+	{
+		"Name": "Weapons",
+		"Bones": [
+			"j_buki_sebo_l",
+			"j_buki_sebo_r",
+			"j_buki2_kosi_l",
+			"j_buki2_kosi_r",
+			"j_buki_kosi_l",
+			"j_buki_kosi_r",
+			"n_buki_r",
+			"n_buki_l",
+			"n_buki_tate_l",
+			"n_buki_tate_r"
+		]
+	}
+]
