+++ Ktisis/Interface/Library/Common.cs
@@ -1,4 +1,6 @@
-﻿using System.Numerics;
+﻿using System.IO;
+using System.Numerics;
+using System.Reflection;
 using System.Diagnostics;
 
 using ImGuiNET;
@@ -13,6 +15,20 @@ namespace Ktisis.Interface.Library {
 		// From SimpleTweaks - Thanks Caraxi
 		internal static void OpenBrowser(string url)
 			=> Process.Start(new ProcessStartInfo { FileName = url, UseShellExecute = true });
+
+		// Open assembly file
+		internal static Stream GetAssemblyFile(string path) {
+			var assembly = Assembly.GetExecutingAssembly();
+			string assemblyName = assembly.GetName().Name!;
+
+			path = $"{assemblyName}.{path}";
+
+			Stream? stream = assembly.GetManifestResourceStream(path);
+			if (stream == null)
+				throw new FileNotFoundException(path);
+
+			return stream;
+		}
 
 		// Item tooltip
 		internal static void Tooltip(string text) {
